refactor iterator_ext, and add test
